Remove logging of which SSH key is used for authentication
This is useful info and it will probably return at some point.
However the logger in use doesn't log things immediately, which
doesn't really play well with the data being allocated right before
and deleted right after.

It's not a big deal, but it does add noise to ASAN/Valgrind which
makes it harder to track down actual issues.