Use a StringBuilder instead of a linked list for ErrorOr errors
This makes things a bit less convenient and a tad confusing (storing
the error message in reverse and so on) but it is enough of a
performance improvement to be worth it. The copy/move/delete of a
forward_list were too expensive for an object that is constantly being
copied/moved/deleted. The improved cache locality is also nice.